Split up AssertionHandler
