+++ src/Handler/Assertion/AssertionHandler.php
@@ -4,133 +4,33 @@ declare(strict_types=1);
 
 namespace webignition\BasilCompilableSourceFactory\Handler\Assertion;
 
-use webignition\BasilCompilableSource\Block\TryCatch\CatchBlock;
-use webignition\BasilCompilableSource\Block\TryCatch\TryBlock;
-use webignition\BasilCompilableSource\Block\TryCatch\TryCatchBlock;
-use webignition\BasilCompilableSource\Body\Body;
 use webignition\BasilCompilableSource\Body\BodyInterface;
-use webignition\BasilCompilableSource\Expression\CastExpression;
-use webignition\BasilCompilableSource\Expression\CatchExpression;
-use webignition\BasilCompilableSource\Expression\ClassDependency;
-use webignition\BasilCompilableSource\Expression\ClosureExpression;
-use webignition\BasilCompilableSource\Expression\ComparisonExpression;
-use webignition\BasilCompilableSource\Expression\EncapsulatedExpression;
-use webignition\BasilCompilableSource\Expression\ExpressionInterface;
-use webignition\BasilCompilableSource\Expression\LiteralExpression;
-use webignition\BasilCompilableSource\MethodInvocation\MethodInvocation;
-use webignition\BasilCompilableSource\MethodInvocation\ObjectMethodInvocation;
-use webignition\BasilCompilableSource\Expression\ObjectPropertyAccessExpression;
-use webignition\BasilCompilableSource\Statement\AssignmentStatement;
-use webignition\BasilCompilableSource\Statement\Statement;
-use webignition\BasilCompilableSource\Statement\StatementInterface;
-use webignition\BasilCompilableSource\TypeDeclaration\ObjectTypeDeclaration;
-use webignition\BasilCompilableSource\TypeDeclaration\ObjectTypeDeclarationCollection;
-use webignition\BasilCompilableSource\VariableDependency;
-use webignition\BasilCompilableSource\VariableName;
-use webignition\BasilCompilableSourceFactory\AccessorDefaultValueFactory;
-use webignition\BasilCompilableSourceFactory\AssertionMethodInvocationFactory;
-use webignition\BasilCompilableSourceFactory\CallFactory\DomCrawlerNavigatorCallFactory;
-use webignition\BasilCompilableSourceFactory\CallFactory\ElementIdentifierCallFactory;
-use webignition\BasilCompilableSourceFactory\ElementIdentifierSerializer;
 use webignition\BasilCompilableSourceFactory\Exception\UnsupportedContentException;
 use webignition\BasilCompilableSourceFactory\Exception\UnsupportedStatementException;
-use webignition\BasilCompilableSourceFactory\Handler\DomIdentifierHandler;
-use webignition\BasilCompilableSourceFactory\Handler\Value\ScalarValueHandler;
-use webignition\BasilCompilableSourceFactory\ValueAccessorFactory;
-use webignition\BasilCompilableSourceFactory\ValueTypeIdentifier;
-use webignition\BasilCompilableSourceFactory\VariableNames;
-use webignition\BasilDomIdentifierFactory\Factory as DomIdentifierFactory;
-use webignition\BasilIdentifierAnalyser\IdentifierTypeAnalyser;
-use webignition\BasilModels\Action\ActionInterface;
-use webignition\BasilModels\Assertion\Assertion;
 use webignition\BasilModels\Assertion\AssertionInterface;
-use webignition\BasilModels\Assertion\DerivedValueOperationAssertion;
-use webignition\DomElementIdentifier\AttributeIdentifierInterface;
-use webignition\DomElementIdentifier\ElementIdentifier;
-use webignition\DomElementIdentifier\ElementIdentifierInterface;
-use webignition\SymfonyDomCrawlerNavigator\Exception\InvalidLocatorException;
 
 class AssertionHandler
 {
-    public const ASSERT_EQUALS_METHOD = 'assertEquals';
-    public const ASSERT_NOT_EQUALS_METHOD = 'assertNotEquals';
-    public const ASSERT_STRING_CONTAINS_STRING_METHOD = 'assertStringContainsString';
-    public const ASSERT_STRING_NOT_CONTAINS_STRING_METHOD = 'assertStringNotContainsString';
-    public const ASSERT_MATCHES_METHOD = 'assertRegExp';
-    public const ASSERT_TRUE_METHOD = 'assertTrue';
-    public const ASSERT_FALSE_METHOD = 'assertFalse';
-
-    private AccessorDefaultValueFactory $accessorDefaultValueFactory;
-    private AssertionMethodInvocationFactory $assertionMethodInvocationFactory;
-    private DomCrawlerNavigatorCallFactory $domCrawlerNavigatorCallFactory;
-    private DomIdentifierFactory $domIdentifierFactory;
-    private DomIdentifierHandler $domIdentifierHandler;
-    private IdentifierTypeAnalyser $identifierTypeAnalyser;
-    private ScalarValueHandler $scalarValueHandler;
-    private ValueTypeIdentifier $valueTypeIdentifier;
-    private ElementIdentifierCallFactory $elementIdentifierCallFactory;
-    private ElementIdentifierSerializer $elementIdentifierSerializer;
-    private ValueAccessorFactory $valueAccessorFactory;
-
-    private const OPERATOR_TO_ASSERTION_TEMPLATE_MAP = [
-        'includes' => self::ASSERT_STRING_CONTAINS_STRING_METHOD,
-        'excludes' => self::ASSERT_STRING_NOT_CONTAINS_STRING_METHOD,
-        'is' => self::ASSERT_EQUALS_METHOD,
-        'is-not' => self::ASSERT_NOT_EQUALS_METHOD,
-        'matches' => self::ASSERT_MATCHES_METHOD,
-        'exists' => self::ASSERT_TRUE_METHOD,
-        'not-exists' => self::ASSERT_FALSE_METHOD,
-        'is-regexp' => self::ASSERT_FALSE_METHOD,
-    ];
-
-    /**
-     * @var string[]
-     */
-    private array $methodsWithStringArguments = [
-        self::ASSERT_STRING_CONTAINS_STRING_METHOD,
-        self::ASSERT_STRING_NOT_CONTAINS_STRING_METHOD,
-    ];
+    private ComparisonAssertionHandler $comparisonAssertionHandler;
+    private ExistenceAssertionHandler $existenceAssertionHandler;
+    private IsRegExpAssertionHandler $isRegExpAssertionHandler;
 
     public function __construct(
-        AccessorDefaultValueFactory $accessorDefaultValueFactory,
-        AssertionMethodInvocationFactory $assertionMethodInvocationFactory,
-        DomCrawlerNavigatorCallFactory $domCrawlerNavigatorCallFactory,
-        DomIdentifierFactory $domIdentifierFactory,
-        DomIdentifierHandler $domIdentifierHandler,
-        IdentifierTypeAnalyser $identifierTypeAnalyser,
-        ScalarValueHandler $scalarValueHandler,
-        ValueTypeIdentifier $valueTypeIdentifier,
-        ElementIdentifierCallFactory $elementIdentifierCallFactory,
-        ElementIdentifierSerializer $elementIdentifierSerializer,
-        ValueAccessorFactory $valueAccessorFactory
+        ComparisonAssertionHandler $comparisonAssertionHandler,
+        ExistenceAssertionHandler $existenceAssertionHandler,
+        IsRegExpAssertionHandler $isRegExpAssertionHandler
     ) {
-        $this->accessorDefaultValueFactory = $accessorDefaultValueFactory;
-        $this->assertionMethodInvocationFactory = $assertionMethodInvocationFactory;
-        $this->domCrawlerNavigatorCallFactory = $domCrawlerNavigatorCallFactory;
-        $this->domIdentifierFactory = $domIdentifierFactory;
-        $this->domIdentifierHandler = $domIdentifierHandler;
-        $this->identifierTypeAnalyser = $identifierTypeAnalyser;
-        $this->scalarValueHandler = $scalarValueHandler;
-        $this->valueTypeIdentifier = $valueTypeIdentifier;
-        $this->elementIdentifierCallFactory = $elementIdentifierCallFactory;
-        $this->elementIdentifierSerializer = $elementIdentifierSerializer;
-        $this->valueAccessorFactory = $valueAccessorFactory;
+        $this->comparisonAssertionHandler = $comparisonAssertionHandler;
+        $this->existenceAssertionHandler = $existenceAssertionHandler;
+        $this->isRegExpAssertionHandler = $isRegExpAssertionHandler;
     }
 
     public static function createHandler(): AssertionHandler
     {
         return new AssertionHandler(
-            AccessorDefaultValueFactory::createFactory(),
-            AssertionMethodInvocationFactory::createFactory(),
-            DomCrawlerNavigatorCallFactory::createFactory(),
-            DomIdentifierFactory::createFactory(),
-            DomIdentifierHandler::createHandler(),
-            IdentifierTypeAnalyser::create(),
-            ScalarValueHandler::createHandler(),
-            new ValueTypeIdentifier(),
-            ElementIdentifierCallFactory::createFactory(),
-            ElementIdentifierSerializer::createSerializer(),
-            ValueAccessorFactory::createFactory()
+            ComparisonAssertionHandler::createHandler(),
+            ExistenceAssertionHandler::createHandler(),
+            IsRegExpAssertionHandler::createHandler()
         );
     }
 
@@ -145,14 +45,14 @@ class AssertionHandler
     {
         try {
             if ($assertion->isComparison()) {
-                return $this->handleComparisonAssertion($assertion);
+                return $this->comparisonAssertionHandler->handle($assertion);
             } else {
                 if (in_array($assertion->getOperator(), ['exists', 'not-exists'])) {
-                    return $this->handleExistenceAssertion($assertion);
+                    return $this->existenceAssertionHandler->handle($assertion);
                 }
 
                 if ('is-regexp' === $assertion->getOperator()) {
-                    return $this->handleIsRegExpAssertion($assertion);
+                    return $this->isRegExpAssertionHandler->handle($assertion);
                 }
             }
         } catch (UnsupportedContentException $previous) {
@@ -160,437 +60,5 @@ class AssertionHandler
         }
 
         throw new UnsupportedStatementException($assertion);
-    }
-
-    /**
-     * @param AssertionInterface $assertion
-     *
-     * @return BodyInterface
-     *
-     * @throws UnsupportedContentException
-     */
-    private function handleExistenceAssertion(AssertionInterface $assertion): BodyInterface
-    {
-        $identifier = $assertion->getIdentifier();
-
-        $assertionStatement = $this->createAssertionStatement($assertion, [
-            $this->createGetBooleanExaminedValueInvocation()
-        ]);
-
-        if ($this->valueTypeIdentifier->isScalarValue($identifier)) {
-            return $this->handleScalarExistenceAssertion($assertion);
-        }
-
-        if ($this->identifierTypeAnalyser->isDomOrDescendantDomIdentifier($identifier)) {
-            $domIdentifier = $this->domIdentifierFactory->createFromIdentifierString($identifier);
-            if (null === $domIdentifier) {
-                throw new UnsupportedContentException(UnsupportedContentException::TYPE_IDENTIFIER, $identifier);
-            }
-
-            $serializedElementIdentifier = $this->elementIdentifierSerializer->serialize($domIdentifier);
-            $elementIdentifierExpression = $this->elementIdentifierCallFactory->createConstructorCall(
-                $serializedElementIdentifier
-            );
-
-            $examinedElementIdentifierPlaceholder = new ObjectPropertyAccessExpression(
-                new VariableDependency(VariableNames::PHPUNIT_TEST_CASE),
-                'examinedElementIdentifier'
-            );
-
-            $domNavigatorCrawlerCall = $this->createExistenceOperationDomCrawlerNavigatorCall(
-                $domIdentifier,
-                $assertion,
-                $examinedElementIdentifierPlaceholder
-            );
-
-            $elementSetBooleanExaminedValueInvocation = $this->createSetBooleanExaminedValueInvocation(
-                [
-                    $domNavigatorCrawlerCall
-                ],
-                ObjectMethodInvocation::ARGUMENT_FORMAT_STACKED
-            );
-
-            if (!$domIdentifier instanceof AttributeIdentifierInterface) {
-                return new Body([
-                    new AssignmentStatement(
-                        $examinedElementIdentifierPlaceholder,
-                        $elementIdentifierExpression
-                    ),
-                    $this->createNavigatorHasCallTryCatchBlock($elementSetBooleanExaminedValueInvocation),
-                    $assertionStatement,
-                ]);
-            }
-
-            $elementIdentifierString = (string) ElementIdentifier::fromAttributeIdentifier($domIdentifier);
-            $elementExistsAssertion = new Assertion(
-                $elementIdentifierString . ' exists',
-                $elementIdentifierString,
-                'exists'
-            );
-
-            $attributeNullComparisonExpression = $this->createNullComparisonExpression(
-                $this->domIdentifierHandler->handleAttributeValue(
-                    $this->elementIdentifierSerializer->serialize($domIdentifier),
-                    $domIdentifier->getAttributeName()
-                )
-            );
-
-            $attributeSetBooleanExaminedValueInvocation = $this->createSetBooleanExaminedValueInvocation([
-                new ComparisonExpression(
-                    new EncapsulatedExpression($attributeNullComparisonExpression),
-                    new LiteralExpression('null'),
-                    '!=='
-                ),
-            ]);
-
-            return new Body([
-                new AssignmentStatement(
-                    $examinedElementIdentifierPlaceholder,
-                    $elementIdentifierExpression
-                ),
-                $this->createNavigatorHasCallTryCatchBlock($elementSetBooleanExaminedValueInvocation),
-                $this->createAssertionStatement($elementExistsAssertion, [
-                    $this->createGetBooleanExaminedValueInvocation()
-                ]),
-                new Statement($attributeSetBooleanExaminedValueInvocation),
-                $assertionStatement,
-            ]);
-        }
-
-        throw new UnsupportedContentException(UnsupportedContentException::TYPE_IDENTIFIER, $identifier);
-    }
-
-    /**
-     * @param AssertionInterface $assertion
-     *
-     * @return BodyInterface
-     *
-     * @throws UnsupportedContentException
-     */
-    private function handleScalarExistenceAssertion(AssertionInterface $assertion): BodyInterface
-    {
-        $nullComparisonExpression = $this->createNullComparisonExpression(
-            $this->scalarValueHandler->handle($assertion->getIdentifier())
-        );
-
-        $setBooleanExaminedValueInvocation = $this->createSetBooleanExaminedValueInvocation([
-            new ComparisonExpression(
-                new EncapsulatedExpression($nullComparisonExpression),
-                new LiteralExpression('null'),
-                '!=='
-            ),
-        ]);
-
-        $assertionStatement = $this->createAssertionStatement($assertion, [
-            $this->createGetBooleanExaminedValueInvocation()
-        ]);
-
-        return new Body([
-            new Statement($setBooleanExaminedValueInvocation),
-            $assertionStatement,
-        ]);
-    }
-
-    private function createNullComparisonExpression(ExpressionInterface $leftHandSide): ExpressionInterface
-    {
-        return new ComparisonExpression($leftHandSide, new LiteralExpression('null'), '??');
-    }
-
-    /**
-     * @param ExpressionInterface[] $arguments
-     * @param string $argumentFormat
-     *
-     * @return ExpressionInterface
-     */
-    private function createSetBooleanExaminedValueInvocation(
-        array $arguments,
-        string $argumentFormat = ObjectMethodInvocation::ARGUMENT_FORMAT_INLINE
-    ): ExpressionInterface {
-        return $this->createPhpUnitTestCaseObjectMethodInvocation(
-            'setBooleanExaminedValue',
-            $arguments,
-            $argumentFormat
-        );
-    }
-
-    /**
-     * @param ExpressionInterface[] $arguments
-     * @param string $argumentFormat
-     *
-     * @return ExpressionInterface
-     */
-    private function createSetBooleanExpectedValueInvocation(
-        array $arguments,
-        string $argumentFormat = ObjectMethodInvocation::ARGUMENT_FORMAT_INLINE
-    ): ExpressionInterface {
-        return $this->createPhpUnitTestCaseObjectMethodInvocation(
-            'setBooleanExpectedValue',
-            $arguments,
-            $argumentFormat
-        );
-    }
-
-    private function createGetBooleanExaminedValueInvocation(): ExpressionInterface
-    {
-        return $this->createPhpUnitTestCaseObjectMethodInvocation('getBooleanExaminedValue');
-    }
-
-    private function createGetBooleanExpectedValueInvocation(): ExpressionInterface
-    {
-        return $this->createPhpUnitTestCaseObjectMethodInvocation('getBooleanExpectedValue');
-    }
-
-    /**
-     * @param ExpressionInterface[] $arguments
-     *
-     * @return ExpressionInterface
-     */
-    private function createSetExaminedValueInvocation(array $arguments): ExpressionInterface
-    {
-        return $this->createPhpUnitTestCaseObjectMethodInvocation(
-            'setExaminedValue',
-            $arguments
-        );
-    }
-
-    private function createGetExaminedValueInvocation(): ExpressionInterface
-    {
-        return $this->createPhpUnitTestCaseObjectMethodInvocation('getExaminedValue');
-    }
-
-    private function createGetExpectedValueInvocation(): ExpressionInterface
-    {
-        return $this->createPhpUnitTestCaseObjectMethodInvocation('getExpectedValue');
-    }
-
-    /**
-     * @param ExpressionInterface[] $arguments
-     *
-     * @return ExpressionInterface
-     */
-    private function createSetExpectedValueInvocation(array $arguments): ExpressionInterface
-    {
-        return $this->createPhpUnitTestCaseObjectMethodInvocation(
-            'setExpectedValue',
-            $arguments
-        );
-    }
-
-    /**
-     * @param string $methodName
-     * @param ExpressionInterface[] $arguments
-     * @param string $argumentFormat
-     *
-     * @return ExpressionInterface
-     */
-    private function createPhpUnitTestCaseObjectMethodInvocation(
-        string $methodName,
-        array $arguments = [],
-        string $argumentFormat = ObjectMethodInvocation::ARGUMENT_FORMAT_INLINE
-    ): ExpressionInterface {
-        return new ObjectMethodInvocation(
-            new VariableDependency(VariableNames::PHPUNIT_TEST_CASE),
-            $methodName,
-            $arguments,
-            $argumentFormat
-        );
-    }
-
-    private function createExistenceOperationDomCrawlerNavigatorCall(
-        ElementIdentifierInterface $domIdentifier,
-        AssertionInterface $assertion,
-        ObjectPropertyAccessExpression $expression
-    ): ExpressionInterface {
-        $isAttributeIdentifier = $domIdentifier instanceof AttributeIdentifierInterface;
-        $isDerivedFromInteractionAction = false;
-
-        if ($assertion instanceof DerivedValueOperationAssertion) {
-            $sourceStatement = $assertion->getSourceStatement();
-
-            $isDerivedFromInteractionAction =
-                $sourceStatement instanceof ActionInterface && $sourceStatement->isInteraction();
-        }
-
-        return $isAttributeIdentifier || $isDerivedFromInteractionAction
-                ? $this->domCrawlerNavigatorCallFactory->createHasOneCall($expression)
-                : $this->domCrawlerNavigatorCallFactory->createHasCall($expression);
-    }
-
-    /**
-     * @param AssertionInterface $assertion
-     *
-     * @return BodyInterface
-     *
-     * @throws UnsupportedContentException
-     */
-    private function handleComparisonAssertion(AssertionInterface $assertion): BodyInterface
-    {
-        $assertionMethod = self::OPERATOR_TO_ASSERTION_TEMPLATE_MAP[$assertion->getOperator()];
-
-        $examinedAccessor = $this->createValueAccessor($assertion->getIdentifier());
-        $expectedAccessor = $this->createValueAccessor($assertion->getValue());
-
-        $assertionArguments = [
-            $this->createGetExpectedValueInvocation(),
-            $this->createGetExaminedValueInvocation(),
-        ];
-
-        $isStringArgumentAssertionMethod = in_array($assertionMethod, $this->methodsWithStringArguments);
-        if ($isStringArgumentAssertionMethod) {
-            array_walk($assertionArguments, function (ExpressionInterface &$expression) {
-                $expression = new CastExpression($expression, 'string');
-            });
-        }
-
-        return new Body([
-            new Statement($this->createSetExpectedValueInvocation([$expectedAccessor])),
-            new Statement($this->createSetExaminedValueInvocation([$examinedAccessor])),
-            $this->createAssertionStatement($assertion, $assertionArguments),
-        ]);
-    }
-
-    /**
-     * @param AssertionInterface $assertion
-     *
-     * @return BodyInterface
-     *
-     * @throws UnsupportedContentException
-     */
-    private function handleIsRegExpAssertion(AssertionInterface $assertion): BodyInterface
-    {
-        $identifier = $assertion->getIdentifier();
-
-        if ($this->valueTypeIdentifier->isScalarValue($identifier)) {
-            $examinedAccessor = new LiteralExpression($identifier);
-
-            return $this->createIsRegExpAssertionBody($examinedAccessor, $assertion);
-        }
-
-        if ($this->identifierTypeAnalyser->isDomOrDescendantDomIdentifier($identifier)) {
-            $domIdentifier = $this->domIdentifierFactory->createFromIdentifierString($identifier);
-            if (null === $domIdentifier) {
-                throw new UnsupportedContentException(UnsupportedContentException::TYPE_IDENTIFIER, $identifier);
-            }
-
-            $examinedAccessor = $this->createValueAccessor($assertion->getIdentifier());
-
-            return $this->createIsRegExpAssertionBody($examinedAccessor, $assertion);
-        }
-
-        throw new UnsupportedContentException(UnsupportedContentException::TYPE_IDENTIFIER, $identifier);
-    }
-
-    private function createIsRegExpAssertionBody(
-        ExpressionInterface $examinedAccessor,
-        AssertionInterface $assertion
-    ): BodyInterface {
-        $pregMatchInvocation = new MethodInvocation(
-            'preg_match',
-            [
-                $this->createGetExaminedValueInvocation(),
-                new LiteralExpression('null'),
-            ]
-        );
-        $pregMatchInvocation->enableErrorSuppression();
-
-        $identityComparison = new ComparisonExpression(
-            $pregMatchInvocation,
-            new LiteralExpression('false'),
-            '==='
-        );
-
-        return new Body([
-            new Statement($this->createSetExaminedValueInvocation([
-                $examinedAccessor
-            ])),
-            new Statement($this->createSetBooleanExpectedValueInvocation(
-                [
-                    $identityComparison
-                ],
-                MethodInvocation::ARGUMENT_FORMAT_STACKED
-            )),
-            $this->createAssertionStatement($assertion, [
-                $this->createGetBooleanExpectedValueInvocation()
-            ]),
-        ]);
-    }
-
-    /**
-     * @param string $value
-     *
-     * @return ExpressionInterface
-     *
-     * @throws UnsupportedContentException
-     */
-    private function createValueAccessor(string $value): ExpressionInterface
-    {
-        $accessor = $this->valueAccessorFactory->create($value);
-
-        if (!$accessor instanceof ClosureExpression) {
-            $defaultValue = $this->accessorDefaultValueFactory->createString($value) ?? 'null';
-
-            $accessor = new ComparisonExpression(
-                $accessor,
-                new LiteralExpression($defaultValue),
-                '??'
-            );
-        }
-
-        return $accessor;
-    }
-
-    /**
-     * @param AssertionInterface $assertion
-     * @param ExpressionInterface[] $arguments
-     *
-     * @return StatementInterface
-     */
-    private function createAssertionStatement(AssertionInterface $assertion, array $arguments): StatementInterface
-    {
-        return new Statement(
-            $this->assertionMethodInvocationFactory->create(
-                self::OPERATOR_TO_ASSERTION_TEMPLATE_MAP[$assertion->getOperator()],
-                $arguments
-            )
-        );
-    }
-
-    private function createNavigatorHasCallTryCatchBlock(
-        ExpressionInterface $elementSetBooleanExaminedValueInvocation
-    ): TryCatchBlock {
-        return new TryCatchBlock(
-            new TryBlock(
-                new Body([
-                    new Statement($elementSetBooleanExaminedValueInvocation)
-                ])
-            ),
-            new CatchBlock(
-                new CatchExpression(
-                    new ObjectTypeDeclarationCollection([
-                        new ObjectTypeDeclaration(new ClassDependency(InvalidLocatorException::class))
-                    ])
-                ),
-                new Body([
-                    new Statement(
-                        new ObjectMethodInvocation(
-                            new VariableDependency(VariableNames::PHPUNIT_TEST_CASE),
-                            'setLastException',
-                            [
-                                new VariableName('exception')
-                            ]
-                        )
-                    ),
-                    new Statement(
-                        new ObjectMethodInvocation(
-                            new VariableDependency(VariableNames::PHPUNIT_TEST_CASE),
-                            'fail',
-                            [
-                                new LiteralExpression('"Invalid locator"'),
-                            ]
-                        )
-                    ),
-                ])
-            )
-        );
     }
 }
